Remove check to allow user's current task to be returned next
If a task is locked by a user, it means the user was currently
(or recently) working on it. Change getRandomTasks to exclude
these so that we are always fetching brand new tasks the user
hasn't seen before. There is a situation where a user is "skipping"
a task that has been skipped before and that task continues to
be returned as the next task.